Split up the V2 files more and made some performance enhancements.
